requirements.txt
# Core dependencies
streamlit==1.49.0
pandas==2.2.3
numpy==2.0.2
Pillow==10.4.0

# AI/ML dependencies
openai==1.108.0
torch==2.6.0
torchvision==0.21.0
transformers==4.38.0
sentence-transformers==2.5.0

# Vector database
chromadb==1.1.0

# Image processing and computer vision
opencv-python==4.10.0.84
scikit-image==0.22.0
git+https://github.com/openai/CLIP.git

# Data handling
kaggle==1.6.6
python-dotenv==1.0.1
tqdm==4.66.3

# Additional utilities
requests==2.32.4
matplotlib==3.10.6
seaborn==0.13.2
plotly==5.19.0
